plant_weather_app/plant_weather_app/requirements.txt
Flask
requests
Pillow
numpy
# Optional (only if you use a Keras model): tensorflow
